backend/requirements/base.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
